AIO fixes / improvements
Complete AIO requests in process instead of signal context.  Keep a list
of pending and a list of completed AIO requests.  Wait for all AIO
requests to be completed before shutting down.

When an AIO request to release a lockspace completes and the lockspace
is still there, more references on the lockspace must still be held.
Repeat the request.